Projects/requirements.txt
selenium
webdriver-manager
pandas
pytest
openpyxl
allure-pytest
configparser